Complete graph and solver code.
Initial run showed that the vertices being visited were getting very
long because of applications of rule 2. I modified the implementation to
store the vertices to visit in a PriorityQueue, with the priority being
the length of the string represented by the vertex. After letting the
code run for <1min, and checking that only novel vertices were being
visited and that the PQ was working as intended, I checked on Wikipedia
what the answer to the problem was.

Fix GH-#3